extract mock url bar
